# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Cenos API
  description: |
    API do ecossistema Cenos para gestão de agentes de campo (leituristas)
    das concessionárias de energia do Piauí (PI) e Maranhão (MA).

    ## Autenticação

    A API suporta 5 métodos de autenticação:

    1. **JWT Bearer** — Painel administrativo (`Authorization: Bearer <token>`)
    2. **Telegram TMA** — App do agente (`X-Telegram-Init-Data`)
    3. **Token Simples** — Consultas externas/BI (`?token=API_TOKEN`)

    Consulte a documentação detalhada em `docs/` para cada módulo.
  version: "2.0.0"
  contact:
    name: Cenos Suporte
    url: https://cenos.com.br
servers:
  - url: https://api.izi.tec.br
    description: Produção
  - url: http://localhost:3040
    description: Desenvolvimento
...
paths:
...
  /admin/badge/{id}:
    get:
      tags: [Admin, Badges]
      summary: Detalhes da badge
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Badge
    put:
      tags: [Admin, Badges]
      summary: Atualiza badge
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                image_url:
                  type: string
      responses:
        "200":
          description: Badge atualizada
    delete:
      tags: [Admin, Badges]
      summary: Remove badge
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Badge removida

  # ─── ADMIN: USER BADGES ──────────────────────────────────────────────
  /admin/user-badges/{id}:
    get:
      tags: [Admin, Badges]
      summary: Badges do agente
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Badges do agente
  /admin/user-badges/{id}/add:
    post:
      tags: [Admin, Badges]
      summary: Atribui badge ao agente
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                badgeId:
                  type: integer
      responses:
        "200":
          description: Badge atribuída
  /admin/user-badges/{id}/remove:
    post:
      tags: [Admin, Badges]
      summary: Revoga badge do agente
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                badgeId:
                  type: integer
      responses:
        "200":
          description: Badge revogada

  # ─── ADMIN: CENEDUC ──────────────────────────────────────────────────
...
  /admin/tracking/agents:
    get:
      tags: [Admin, Tracking]
      summary: Última posição de todos agentes
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Lista de agentes com posição
  /admin/tracking/agent/{id}/trail:
    get:
      tags: [Admin, Tracking]
      summary: Trilha histórica do agente
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Coordenadas da trilha
  /admin/tracking/speed_violations:
    get:
      tags: [Admin, Tracking]
      summary: Violações de velocidade
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Lista de violações
  /admin/tracking/fall_incidents:
    get:
      tags: [Admin, Tracking]
      summary: Incidentes de queda
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Incidentes detectados
  /admin/tracking/fall_incidents/{id}:
    put:
      tags: [Admin, Tracking]
      summary: Atualiza status do incidente
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                notes:
                  type: string
      responses:
        "200":
          description: Incidente atualizado

  # ─── ADMIN: FORMS ────────────────────────────────────────────────────
...
  /admin/user/users/{id}:
    get:
      tags: [Admin]
      summary: Detalhes de um usuário
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Usuário não encontrado
    put:
      tags: [Admin]
      summary: Atualizar dados do usuário
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                nome:
                  type: string
                role:
                  type: string
                estado:
                  type: string
      responses:
        "200":
          description: Usuário atualizado
        "404":
          description: Não encontrado
    delete:
      tags: [Admin]
      summary: Remover usuário
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário removido
        "404":
          description: Não encontrado
...
components:
...
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticação administrativa via JWT.
        Header: `Authorization: Bearer <token>`
        O token é obtido via `POST /admin/user/login`.
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: |
        Autenticação do agente via Telegram Mini App.
        Header: `X-Telegram-Init-Data: <initData>`
        O initData é validado via Telegram TMA.
    ApiTokenAuth:
      type: apiKey
      in: query
      name: token
      description: |
        Token simples para consultas externas.
        Query param: `?token=API_TOKEN`
  schemas:
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nome:
          type: string
        role:
          type: string
        estado:
          type: string
...